fix(b7-eval): cast merged ckpt to bfloat16 + use TP=8 for 35B on A100-40GB
Previous 35B path saved 140GB FP32 (from_config defaults to FP32),
then vllm OOM'd loading on A100:40GB with TP=4.  Cast to BF16 at merge
time → 70GB on disk; bump TP from 4→8 to fully shard across 8 GPUs.

- merge_fsdp_checkpoint.py: pass torch_dtype=torch.bfloat16 to
  model_class.from_config + model.to(torch.bfloat16) safety-net cast
  after load_state_dict.
- witness-eval-arc-score.yaml: 35B TP_SIZE 4→8, MEM_UTIL 0.70→0.85.

+++ examples/train_integrations/witness/merge_fsdp_checkpoint.py
@@ -326,10 +326,12 @@ def main():
         model_class = AutoModelForCausalLM
 
     with torch.device("meta"):
-        model = model_class.from_config(config, trust_remote_code=True)
+        model = model_class.from_config(
+            config, trust_remote_code=True, torch_dtype=torch.bfloat16,
+        )
 
-    # to_empty allocates real (empty) storage on cpu, then load_state_dict fills it
     model.to_empty(device="cpu")
+    model = model.to(torch.bfloat16)
 
     # Pre-filter: drop ckpt keys whose shape differs from the model's
     # current parameter — strict=False allows missing/unexpected keys, but
